fix: 编译错误 - SubjectCountItem 访问权限 + AnswerFloatWindow 类型不匹配
1. QuestionBankManageScreen.kt:
   - SubjectCountItem 数据类移到 composable 函数外部，解决访问权限错误

2. AnswerFloatWindow.kt:
   - textAreaWidth 计算: coerceAtLeast(50) → coerceAtLeast(50f) 解决 Float 类型不匹配
   - 添加 .toInt() 转换，满足 StaticLayout.Builder.obtain() 的 Int 参数要求

+++ app/src/main/java/com/smartsearch/app/feature/search/floatview/AnswerFloatWindow.kt
@@ -232,7 +232,8 @@ class AnswerFloatWindow(private val context: Context) : View(context) {
         // 文本区域宽度 = 窗口宽度 - 左侧关闭按钮区 - 右侧选区按钮区 - 内边距
         val closeArea = padding + closeButtonRadius * 2 + padding
         val selectArea = padding + selectAreaBtnWidth + padding
-        val textAreaWidth = (currentWidth - closeArea - selectArea - padding * 2).coerceAtLeast(50)
+        val textAreaWidth = (currentWidth - closeArea - selectArea - padding * 2).coerceAtLeast(50f)
+            .toInt()
 
         val displayText = if (answerText.isNotEmpty()) answerText else explanationText
